fix(Avatar): implement history with the theme of presente
