+++ src/components/data/Avatar/stories/Avatar.Default.stories.tsx
@@ -1,4 +1,8 @@
-import { Avatar } from "../index";
+import { ThemeProvider } from "styled-components";
+
+import { IAvatarProps } from "../interfaces/Avatar.interface";
+import { presente } from "@src/shared/themes/presente";
+import { Avatar } from "..";
 
 import { props } from "./props";
 
@@ -10,5 +14,15 @@ const story = {
 };
 
 export const Default = () => <Avatar />;
+
+const theme = {
+  ...presente,
+};
+
+export const Themed = (args: IAvatarProps) => (
+  <ThemeProvider theme={theme}>
+    <Avatar {...args} />
+  </ThemeProvider>
+);
 
 export default story;
